requirements.txt
dj-database-url==0.5.0
Django==2.2.3
django-appconf==1.0.3
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-decouple==2.1
django-heroku==0.3.1
django-tempus-dominus==5.1.2.5
gunicorn==20.0.4
Pillow==6.1.0
psycopg2==2.8.4
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
whitenoise==5.0.1
